requirements.txt
absl-py==0.13.0
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
beautifulsoup4==4.6.0
bleach==4.1.0
blinker==1.4
cached-property==1.5.2
cachetools==4.2.2
certifi==2018.4.16
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
cycler==0.10.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
filelock==3.0.12
Flask==1.0
Flask-Cors==3.0.9
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.24
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
idna==2.7
imageio==2.9.0
imageio-ffmpeg==0.4.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==4.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mistune==0.8.4
mutapath==0.16.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
ninja==1.10.2
notebook==6.4.4
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.1
opensimplex==0.3
openslide-python==1.1.2
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.1
pandocfilters==1.5.0
parso==0.8.2
path==15.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycurl==7.19.5
pydeck==0.7.0
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pyspng==0.1.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scipy==1.7.1
Send2Trash==1.8.0
singletons==0.2.5
six==1.15.0
smmap==4.0.0
streamlit==1.0.0
tables==3.6.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tifffile==2021.8.8
toml==0.10.2
toolz==0.11.1
torch==1.7.1
torchaudio==0.9.0
torchvision==0.8.2
tornado==6.1
tqdm==4.62.0
traitlets==5.1.0
typing-extensions==3.7.4.3
tzlocal==3.0
urllib3==1.24.3
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
